// src/app/modules/modulo-sim/models/tx-sim.model.ts
import { TxSIMRespiratorioModel } from './tx-sim-respiratorio.model';
import { TxSIMLesionTimoModel } from './tx-sim-lesion-timo.model';
import { TxSIMLesionesModel } from './tx-sim-lesiones.model';
import { TxSIMLesionBursaModel } from './tx-sim-lesion-bursa.model';
import { TxSIMIndiceBursalModel } from './tx-sim-indice-bursal.model';
import { TxSIMFotosModel } from './tx-sim-fotos.model';
import { TxSIMDigestivoModel } from './tx-sim-digestivo.model';
export class TxSIMModel {
    idSIM: number;
    codigoEmpresa: string;
    codigoPlanta: string;
    edad: number;
    sexo: string;
    zona: string;
    galpon: number;
    nroPollos: number;
    fecRegistro: Date;
    fecHoraRegistro:Date;
    responsableInvetsa: string;
    responsableIncubadora: string;
    relacionAFavorBursa: number;
    relacionAFavorBazo: number;
    relacion1a1: number;
    aparienciaNormal: number;
    aparienciaAnormal: number;
    tamanoNormal: number;
    tamanoAnormal: number;
    observacionInvetsa: string;
    observacionPlanta: string;
    flgCerrado: boolean;
    idUsuarioCierre: number;
    fecCierre: Date;
    emailFrom: string;
    emailTo: string;
    usuarioCierre: string;
    nombreArchivo: string;
    descripcionEmpresa: string;
    descripcionPlanta: string;
    descripcionTipoExplotacion: string;
    descripcionSubTipoExplotacion: string;
    idSubTipoExplotacion: number;
    usuarioCreacion: string;
    flgMigrado: boolean;
    listaTxSIMDigestivos: TxSIMDigestivoModel[];
    listaTxSIMFotos: TxSIMFotosModel[];
    listaTxSIMIndiceBursal: TxSIMIndiceBursalModel[];
    listaTxSIMLesionBursa: TxSIMLesionBursaModel[];
    listaTxSIMLesiones: TxSIMLesionesModel[];
    listaTxSIMLesionTimo: TxSIMLesionTimoModel[];
    listaTxSIMRespiratorio: TxSIMRespiratorioModel[];

    // Auditoria
    regUsuario?: number;
    regEstacion?: string;

    constructor() {
        this.idSIM = 0;
        this.codigoEmpresa = '';
        this.codigoPlanta = '';
        this.edad = 0;
        this.sexo = '';
        this.zona = '';
        this.galpon = 0;
        this.nroPollos = 0;
        this.fecRegistro = null;
        this.fecHoraRegistro = null;
        this.responsableInvetsa = '';
        this.responsableIncubadora = '';
        this.relacionAFavorBursa = 0;
        this.relacionAFavorBazo = 0;
        this.relacion1a1 = 0;
        this.aparienciaNormal = 0;
        this.aparienciaAnormal = 0;
        this.tamanoNormal = 0;
        this.tamanoAnormal = 0;
        this.observacionInvetsa = '';
        this.observacionPlanta = '';
        this.flgCerrado = false;
        this.idUsuarioCierre = 0;
        this.fecCierre = null;
        this.emailFrom = '';
        this.emailTo = '';
        this.usuarioCreacion = '';
        this.nombreArchivo = '';
        this.descripcionEmpresa = '';
        this.descripcionPlanta = '';
        this.descripcionTipoExplotacion = '';
        this.descripcionSubTipoExplotacion = '';
        this.idSubTipoExplotacion = 0;
        this.usuarioCierre = '';
        this.flgMigrado = false;
        this.listaTxSIMDigestivos = [];
        this.listaTxSIMFotos = [];
        this.listaTxSIMIndiceBursal = [];
        this.listaTxSIMLesionBursa = [];
        this.listaTxSIMLesiones = [];
        this.listaTxSIMLesionTimo = [];
        this.listaTxSIMRespiratorio = [];

        this.regUsuario = 0;
        this.regEstacion = '';
    }
}